Fix a few ffmpeg issues

+++ src/stdlib/src/lib.rs
@@ -728,7 +728,9 @@ pub fn convert_integer(s: *const c_char, base: c_int) -> Option<(c_ulong, isize,
     let mut overflowed = false;
 
     loop {
-        let val = unsafe { LOOKUP_TABLE[*s.offset(idx) as usize] };
+        // `-1 as usize` is -1
+        // `-1 as u8 as usize` is 255
+        let val = unsafe { LOOKUP_TABLE[*s.offset(idx) as u8 as usize] };
         if val == -1 || val as c_int >= base {
             break;
         } else {
